+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [unreleased]
 
+## Changed
+
+- The `parse_space_packets` function analysis queue argument is now expected to be filled
+  on the right side.
+
 # [v0.18.0rc1] 2023-09-04
 
 ## Added